chore(scripts): add utility getWorkspacePaths
