+++ scripts/update-versions/default.ts
@@ -3,14 +3,10 @@
 // Updates versions for internal packages `@aws-sdk/*` to `*`
 // in dependencies/devDependencies/peerDependencies
 
-import { readdirSync, readFileSync, writeFileSync } from "fs";
+import { readFileSync, writeFileSync } from "fs";
 import { join } from "path";
 
-const rootDir = join(__dirname, "..", "..");
-const packageJsonPath = join(rootDir, "package.json");
-const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
-
-const packages = packageJson.workspaces.packages;
+import { getWorkspacePaths } from "./getWorkspacePaths";
 
 const replaceInternalDepVersionWithAsterisk = (section: { [key: string]: string }) => {
   for (const [key, value] of Object.entries(section)) {
@@ -20,21 +16,13 @@ const replaceInternalDepVersionWithAsterisk = (section: { [key: string]: string 
   }
 };
 
-packages
-  .map((dir: string) => dir.replace("/*", ""))
-  .flatMap((workspacesDir) =>
-    readdirSync(join(rootDir, workspacesDir), { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => dirent.name)
-      .map((workspaceDir) => join(rootDir, workspacesDir, workspaceDir))
-  )
-  .forEach((workspacePath) => {
-    const packageJsonPath = join(workspacePath, "package.json");
-    const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
+getWorkspacePaths().forEach((workspacePath) => {
+  const packageJsonPath = join(workspacePath, "package.json");
+  const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
 
-    ["dependencies", "devDependencies"]
-      .filter((section) => packageJson[section] !== undefined)
-      .forEach((section) => replaceInternalDepVersionWithAsterisk(packageJson[section]));
+  ["dependencies", "devDependencies"]
+    .filter((section) => packageJson[section] !== undefined)
+    .forEach((section) => replaceInternalDepVersionWithAsterisk(packageJson[section]));
 
-    writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2).concat(`\n`));
-  });
+  writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2).concat(`\n`));
+});
