+++ scripts/update-versions/exact.ts
@@ -3,31 +3,19 @@
 // Updates versions for internal packages `@aws-sdk/*` to exact versions
 // in dependencies/devDependencies/peerDependencies
 
-import { readdirSync, readFileSync, writeFileSync } from "fs";
+import { readFileSync, writeFileSync } from "fs";
 import { basename, join } from "path";
 
-const rootDir = join(__dirname, "..", "..");
-const packageJsonPath = join(rootDir, "package.json");
-const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
+import { getWorkspacePaths } from "./getWorkspacePaths";
 
-const packages = packageJson.workspaces.packages;
-
-const packageNameToExactVersionHash = packages
-  .map((dir) => dir.replace("/*", ""))
-  .flatMap((workspacesDir) =>
-    readdirSync(join(rootDir, workspacesDir), { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => dirent.name)
-      .map((workspaceDir) => join(rootDir, workspacesDir, workspaceDir))
-  )
-  .reduce((acc, workspacePath) => {
-    const packageJsonPath = join(workspacePath, "package.json");
-    const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
-    return {
-      ...acc,
-      [`@aws-sdk/${basename(workspacePath)}`]: packageJson.version,
-    };
-  }, {});
+const packageNameToExactVersionHash = getWorkspacePaths().reduce((acc, workspacePath) => {
+  const packageJsonPath = join(workspacePath, "package.json");
+  const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
+  return {
+    ...acc,
+    [`@aws-sdk/${basename(workspacePath)}`]: packageJson.version,
+  };
+}, {});
 
 const replaceInternalDepVersionWithExact = (section: { [key: string]: string }) => {
   for (const [key, value] of Object.entries(section)) {
@@ -37,21 +25,13 @@ const replaceInternalDepVersionWithExact = (section: { [key: string]: string }) 
   }
 };
 
-packages
-  .map((dir) => dir.replace("/*", ""))
-  .flatMap((workspacesDir) =>
-    readdirSync(join(rootDir, workspacesDir), { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => dirent.name)
-      .map((workspaceDir) => join(rootDir, workspacesDir, workspaceDir))
-  )
-  .forEach((workspacePath) => {
-    const packageJsonPath = join(workspacePath, "package.json");
-    const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
+getWorkspacePaths().forEach((workspacePath) => {
+  const packageJsonPath = join(workspacePath, "package.json");
+  const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
 
-    ["dependencies", "devDependencies"]
-      .filter((section) => packageJson[section] !== undefined)
-      .forEach((section) => replaceInternalDepVersionWithExact(packageJson[section]));
+  ["dependencies", "devDependencies"]
+    .filter((section) => packageJson[section] !== undefined)
+    .forEach((section) => replaceInternalDepVersionWithExact(packageJson[section]));
 
-    writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2).concat(`\n`));
-  });
+  writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2).concat(`\n`));
+});
